+++ 404.php
@@ -1,18 +1,13 @@
 <?php
 /**
- * The index file for the theme.
+ * The 404 page for the theme.
  */
 
 ?>
 <?php get_header(); ?>
 <div class="template-content">
 	<main id="skip-link-target" class="flow is-layout-constrained has-global-padding">
-		<?php
-		while ( have_posts() ) {
-			the_post();
-			the_content();
-		}
-		?>
+		<?php block_template_part( '404' ); ?>
 	</main>
 </div>
 <?php get_footer(); ?>