+++ functions.php
@@ -7,7 +7,7 @@
 
 namespace Launchpad;
 
-define( 'LAUNCHPAD_THEME_VERSION', '1.0.0-beta.23' );
+define( 'LAUNCHPAD_THEME_VERSION', '1.0.0-beta.24' );
 
 // Require autoloader.
 require __DIR__ . '/vendor/autoload.php';
@@ -22,6 +22,7 @@ require_once __DIR__ . '/functions/compatibility/wp-rocket.php';
 // Other
 require_once __DIR__ . '/functions/acf.php';
 require_once __DIR__ . '/functions/body-class.php';
+require_once __DIR__ . '/functions/content-filters.php';
 require_once __DIR__ . '/functions/custom-post-types.php';
 require_once __DIR__ . '/functions/enqueue-assets.php';
 require_once __DIR__ . '/functions/helpers.php';